# [Core/data] use wait based prefetcher by default