@scv119 2023-04-30
release test result:

1. streaming_data_ingest_benchmark_1tb
[with this pr](https://buildkite.com/ray-project/release-tests-pr/builds/36626#0187c9c5-873e-4b5e-a877-5aed41dbc7ea): success! total time 64.92455434799194
[baseline](https://buildkite.com/ray-project/release-tests-branch/builds/1612#0187c4d4-0ed6-4fc6-ae51-a448c9505889): success! total time 64.79763126373291

2.  pipelined_data_ingest_benchmark_1tb.aws
[with this pr](https://buildkite.com/ray-project/release-tests-pr/builds/36626#0187c9c5-8739-4797-9cd5-4bc7fbf9aa05): 
```
(ConsumingActor pid=186, ip=10.0.58.86) ##### Overall Pipeline Time Breakdown #####                                            
(ConsumingActor pid=186, ip=10.0.58.86) * Time stalled waiting for next datastream: 7.6ms min, 1.2s max, 161.99ms mean, 4.21s total
(PipelineSplitExecutorCoordinator pid=5130) 2023-04-28 14:53:36,146     INFO streaming_executor.py:147 -- Shutting down <StreamingExecutor(Thread-30, stopped daemon 139750702831360)>.
success! total time 45.71649169921875                                                                                                   
(ConsumingActor pid=186, ip=10.0.49.88) Time to read all data 42.747634274999996 seconds [repeated 19x across cluster] (Ray deduplicates logs by default. Set RAY_DEDUP_LOGS=0 to disable log deduplication, or see https://docs.ray.io/en/master/ray-observability/ray-logging.html#log-deduplication for more options.)
(ConsumingActor pid=186, ip=10.0.49.88) P50/P95/Max batch delay (s) 0.009423945000008871 0.13930677815000647 1.6963853930000141 [repeated 19x across cluster]
(ConsumingActor pid=186, ip=10.0.49.88) Num epochs read 1 [repeated 19x across cluster]
(ConsumingActor pid=186, ip=10.0.49.88) Num batches read 1280 [repeated 19x across cluster]
(ConsumingActor pid=186, ip=10.0.58.86)  [repeated 5x across cluster]
(ConsumingActor pid=186, ip=10.0.49.88) Mean throughput 1197.73 MiB/s [repeated 19x across cluster]
(ConsumingActor pid=186, ip=10.0.49.88) Num bytes read 51200.0 MiB [repeated 18x across cluster]
```

[baseline](https://buildkite.com/ray-project/release-tests-branch/builds/1612#0187c4d4-0ed6-4fc6-ae51-a448c9505889):
#### Overall Pipeline Time Breakdown #####                                           
(ConsumingActor pid=962, ip=10.0.44.103) * Time stalled waiting for next datastream: 7.95ms min, 1.21s max, 122.83ms mean, 3.19s total
(ConsumingActor pid=962, ip=10.0.44.103)                                                                                       
success! total time 46.14367604255676                                                                                                   
(ConsumingActor pid=4712) Time to read all data 43.25214273300003 seconds [repeated 19x across cluster] (Ray deduplicates logs by default. Set RAY_DEDUP_LOGS=0 to disable log deduplication, or see https://docs.ray.io/en/master/ray-observability/ray-logging.html#log-deduplication for more options.)
(ConsumingActor pid=4712) P50/P95/Max batch delay (s) 0.009616004500003328 0.14022980685003572 1.838354100999993 [repeated 19x across cluster]
(ConsumingActor pid=4712) Num epochs read 1 [repeated 19x across cluster]
(ConsumingActor pid=4712) Num batches read 1280 [repeated 19x across cluster]
(ConsumingActor pid=4712) Num bytes read 51200.0 MiB [repeated 19x across cluster]
(ConsumingActor pid=4712) Mean throughput 1183.76 MiB/s [repeated 19x across cluster]
```